test: move test to corresponding files
